Add configuration to enable SSL
I recommend not using this feature to serve directly
call the API of the server but instead use a reverse proxy
to handle HTTPS for you. The implementation of OpenSSL API
in vibe doesn't appear to be fully stable.

Nonetheless this feature is now here and can be used for e.g.
server-to-server communication or if you really want to directly
accessing the API from the browser.

Closes #20