Overwrite background alpha (pal color 0) with 0x00
Usually the alpha value in the palette for the background
is 0x00 but can happen (for example in custom sprites) that
instead its 0xFF.

We overwrite the alpha value now to always be transparent.
Might need to revert this change and instead the sprites need
adjustments. For example I remember that quite some .spr images
of item icons used to be broken and even rendered with an opaque
background in-game.

References issue #46.